test(refactor): organize test dir
