+++ tests/unit/test_config.py
@@ -7,7 +7,7 @@ import yaml
 
 from vminfo_parser.config import Config
 
-from . import const as test_const
+from .. import const as test_const
 
 
 def test_yaml_load(config_dict: dict, yaml_config: str) -> None:
